+++ src/models/error.rs
@@ -2,19 +2,21 @@
 
 use std::{error::Error, fmt};
 
+use actix_web::{http::StatusCode, HttpResponse, ResponseError};
+use deadpool_redis::PoolError;
 use redis::RedisError;
 
 /// Result type used throughout the application.
 pub type OcyResult<T> = Result<T, OcyError>;
 
 /// Error type used throughout the application.
-#[derive(Debug, PartialEq)]
+#[derive(Debug)]
 pub enum OcyError {
     /// Error occurred during interaction with Redis.
     Redis(RedisError),
 
     /// Error occurred while trying to get a pooled DB connection to Redis.
-    RedisConnection(String),
+    RedisConnection(PoolError),
 
     /// Operation attempted on queue that does not exist.
     NoSuchQueue(String),
@@ -47,6 +49,22 @@ impl OcyError {
     }
 }
 
+impl PartialEq for OcyError {
+    fn eq(&self, other: &Self) -> bool {
+        match (self, other) {
+            (Self::Redis(l0), Self::Redis(r0)) => l0 == r0,
+            (Self::RedisConnection(l0), Self::RedisConnection(r0)) => l0.to_string() == r0.to_string(),
+            (Self::NoSuchQueue(l0), Self::NoSuchQueue(r0)) => l0 == r0,
+            (Self::NoSuchJob(l0), Self::NoSuchJob(r0)) => l0 == r0,
+            (Self::BadRequest(l0), Self::BadRequest(r0)) => l0 == r0,
+            (Self::Conflict(l0), Self::Conflict(r0)) => l0 == r0,
+            (Self::Internal(l0), Self::Internal(r0)) => l0 == r0,
+            (Self::ParseError(l0), Self::ParseError(r0)) => l0 == r0,
+            _ => false,
+        }
+    }
+}
+
 impl From<RedisError> for OcyError {
     fn from(err: RedisError) -> Self {
         OcyError::Redis(err)
@@ -59,9 +77,9 @@ impl From<serde_json::Error> for OcyError {
     }
 }
 
-impl From<OcyError> for actix_web::body::Body {
-    fn from(o: OcyError) -> actix_web::body::Body {
-        o.to_string().into_bytes().into()
+impl From<deadpool_redis::PoolError> for OcyError {
+    fn from(err: deadpool_redis::PoolError) -> Self {
+        OcyError::RedisConnection(err)
     }
 }
 
@@ -85,6 +103,25 @@ impl Error for OcyError {
         match self {
             OcyError::Redis(err) => err.source(),
             _ => None,
+        }
+    }
+}
+
+impl ResponseError for OcyError {
+    fn error_response(&self) -> HttpResponse {
+        actix_web::HttpResponseBuilder::new(self.status_code())
+            .insert_header((actix_web::http::header::CONTENT_TYPE, "text/html; charset=utf-8"))
+            .body(self.to_string())
+    }
+
+    fn status_code(&self) -> StatusCode {
+        match self {
+            // TODO: check that this is the right response for ParseError
+            OcyError::Redis(_) | OcyError::Internal(_) | OcyError::ParseError(_) => StatusCode::INTERNAL_SERVER_ERROR,
+            OcyError::RedisConnection(_) => StatusCode::SERVICE_UNAVAILABLE,
+            OcyError::NoSuchQueue(_) | OcyError::NoSuchJob(_) => StatusCode::NOT_FOUND,
+            OcyError::BadRequest(_) => StatusCode::BAD_REQUEST,
+            OcyError::Conflict(_) => StatusCode::CONFLICT,
         }
     }
 }